requirements.txt
blinker==1.4
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
py-mini-racer==0.4.0
pycryptodome==3.9.9
PyQRCode==1.2.1
python-dotenv==0.15.0
requests==2.25.0
sq-native==1.0.5.0.0
SQLAlchemy==1.3.20
sqreen==1.25.1
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
psycopg2==2.8.6
flask-talisman==0.7.0
